+++ apidaora/route/decorator.py
@@ -1,10 +1,9 @@
-from typing import Any, Callable, Optional, Union
+from typing import Any, Callable, Union
 
 from ..asgi.router import Controller
 from ..controllers.background_task import BackgroundTask, make_background_task
 from ..exceptions import InvalidRouteArgumentsError, MethodNotFoundError
 from ..method import MethodType
-from ..middlewares import Middlewares
 from .factory import make_route
 
 
@@ -28,25 +27,35 @@ class _RouteDecorator:
                 raise MethodNotFoundError(attr_name)
 
         def decorator(
-            path_pattern: str,
-            middlewares: Optional[Middlewares] = None,
-            options: bool = False,
-            **kwargs: Any,
+            path_pattern: str, **kwargs: Any
         ) -> Callable[[Callable[..., Any]], Union[Controller, BackgroundTask]]:
-            if len(kwargs) > 0 and tuple(kwargs.keys()) != (
-                'tasks_repository',
+            keys = tuple(kwargs.keys())
+            if len(kwargs) > 0 and not any(
+                (
+                    'tasks_repository' in keys,
+                    'single_running' in keys,
+                    'middlewares' in keys,
+                    'options' in keys,
+                )
             ):
                 raise InvalidRouteArgumentsError(kwargs)
 
             def wrapper(
                 controller: Callable[..., Any]
             ) -> Union[Controller, BackgroundTask]:
+                middlewares = kwargs.get('middlewares')
+                options = kwargs.get('options')
+
                 if brackground:
                     tasks_repository = kwargs.get('tasks_repository')
+                    single_running = kwargs.get('single_running')
                     return make_background_task(
                         controller,
                         path_pattern,
                         tasks_repository=tasks_repository,
+                        single_running=single_running,  # type: ignore
+                        middlewares=middlewares,
+                        options=options,  # type: ignore
                     )
 
                 else:
@@ -55,7 +64,7 @@ class _RouteDecorator:
                         MethodType[method],
                         controller,
                         route_middlewares=middlewares,
-                        options=options,
+                        options=options,  # type: ignore
                     )
                     return route.controller