+++ docs/src/background_task_single_running_controller/background_task_single_running_controller.py
@@ -1,0 +1,12 @@
+import time
+
+from apidaora import appdaora, route
+
+
+@route.background('/hello-single', single_running=True)
+def hello_task(name: str) -> str:
+    time.sleep(1)
+    return f'Hello {name}!'
+
+
+app = appdaora(hello_task)