+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Create single running feature for background tasks
+
 ### 0.18.0 - 2020-04-11
 
  - Create prevent request middleware